device_server: pass options to DeviceServer to enable logging (#180)
When the multiprocessing start method is not fork (Windows only has
spawn, and macOS defaults to spawn), then configuration done in the
main process is not copied to the child process, i.e., the actual
device process.  This means that the logging level set on the main
process is not inherited by the DeviceServer process.

This commit passes the device server options to the DeviceServer so
that it can redo the required logging configuration on the new
process.  It does this by passing all options since we expect to
expand those.

This commit also adds dependency on the dataclasses module which means
being dependent on Python 3.7 (which is now almost 3 years old).